admin chart can show various days, and added form to delete reservations

+++ src/main/java/com/caramel/restaurant/model/view/reservation/ReservationDAO.java
@@ -22,5 +22,5 @@ public interface ReservationDAO {
 
 	public void save(Reservation res);
 		
-	public void deleteById(int id);
+	public void deleteById(long id);
 }
